+++ resolve/hykkt/spgemm/SpGEMMCpu.cpp
@@ -18,11 +18,8 @@ namespace ReSolve
      * @param beta[in] - Scalar multiplier for the sum.
      */
     SpGEMMCpu::SpGEMMCpu(real_type alpha, real_type beta)
+      : alpha_(alpha), beta_(beta)
     {
-      alpha_[0] = alpha;
-      alpha_[1] = 0.0;
-      beta_[0] = beta;
-      beta_[1] = 0.0;
       cholmod_start(&Common_);
 
       A_ = nullptr;
@@ -87,7 +84,7 @@ namespace ReSolve
     void SpGEMMCpu::compute()
     {
       cholmod_sparse* C_chol = cholmod_ssmult(B_, A_, 0, 1, 0, &Common_); // B_ and A_ are reversed because cholmod_sparse is a CSC matrix
-      cholmod_sparse* E_chol = cholmod_add(C_chol, D_, alpha_, beta_, 1, 0, &Common_);
+      cholmod_sparse* E_chol = cholmod_add(C_chol, D_, &alpha_, &beta_, 1, 0, &Common_);
 
       int*       chol_row_data = static_cast<int*>(E_chol->p);
       int*       chol_col_data = static_cast<int*>(E_chol->i);