+++ CHANGELOG.md
@@ -2,11 +2,9 @@
 
 ## HyKKT Release changes
 
-- Added classes and tests for permutation, Ruiz scaling, Cholesky factorization, Schur complement conjugate gradient and matrix multiplication and addition.
+- Added classes and tests for HyKKT solver, permutation, Ruiz scaling, Cholesky factorization, Schur complement conjugate gradient and matrix multiplication and addition.
 
 - Changed random number generation int tests to be C++ style and fixed-seed, to avoid random failures.
-
-- Added Schur Complement Conjugate Gradient class.
 
 ## Changes to Re::Solve since release 0.99.2